エフェクトの座標設定の方法の追加

+++ src/Character/CharacterBase.cpp
@@ -191,7 +191,7 @@ void CharacterBase::AttackUpdate()
 //---------------------------------------------------------------------------
 // 当たり判定をとってほしいタイミングを保存するための関数
 //---------------------------------------------------------------------------
-void CharacterBase::SetHitTiming(int attack_frame_start, int attack_frame_end, bool can_hitstop, int attack_num )
+void CharacterBase::SetHitTiming(float attack_frame_start, float attack_frame_end, bool can_hitstop, int attack_num )
 {
 	// 当たり判定をとってほしい最初のフレームの保存
 	m_attack_hit_damage[attack_num]->start_time = attack_frame_start;
